remove containerIndex and add simple variable reassigment
This mapping was unnecessary, the iterations are usually very minimal and we will not have any use for the mapped indexes outside of the function - it is therefore unreasonable to maintain it. We can simply use the index we find on the fly and discard it on the next iteration.