refactor: iterate terminated containers (#5)
* remove usage.md doc

This information is now in the README and no longer required

* update docstrings for functions and structs

* remove containerIndex and add simple variable reassigment

This mapping was unnecessary, the iterations are usually very minimal and we will not have any use for the mapped indexes outside of the function - it is therefore unreasonable to maintain it. We can simply use the index we find on the fly and discard it on the next iteration.